Setup to creating procedural Voronoi noise
Really went on a journey with this one as I realized I didn't completely understand how grass should be using noise distribution. Then every separate thing I tried to accomplish would not work. URP will not let me draw directly to the screen and I did so much to get around that and view the generated texture in the editor. Might switch to using a compute shader for this later, I need to see how performant generating noise in a surface shader is compared to it.