Grass now supports lighting and recieves shadows.
It doesn't cast shadows. That part wasn't working (probably instancing's fault) and it seemed like it was going to be a huge performance hit to do the vertex shader calculations a second time.

This took so long to figure out because URP in hlsl has no good documentation. This project is going to drive me crazy, it's only a matter of time.