Transfered shader logic for grass to main scene
Also had to learn how to use GraphicsBuffers and Graphics.RenderPrimitivesIndexedIndirect because that seems to be the modern version that is intended in URP. The amount of research that went into this is way more than I expected for simply transferring logic over.

Buffers are not working in the grass scene currently, I need to find a  way to fix it.