Addressables + Variant Collections
A lot of this was a workflow change as I realized what would and wouldn't work as I went from the editor to a build. I wanted a build so that I could actually use profilers to see shader performance, and that led to a butterfly effect of things I had to do.

I also partially split up the logic for chunks here so we're currently not redoing the compute shader every frame, only the first frame. Performance is showing that the current grass shader is actually the thing tanking performance, so I wanted to focus on and profile that first before actually working on the chunk scheduler