Optimization for bezier calculations
Had to derive a bunch more things. I expect this will make things a lot more efficient on the computers of people with good gpu's, which unfortunately I do not have with my 8 gb vram.

My bottleneck seems to be the fragment part of the shader where I calculate pbr for each grass blade. This is somewhat necessary for good lighting. I can cut down on chunks a lot by shrinking the map, which condenses the blades together to make the field look fuller. I can do 16x16 chunks instead of 64x64 that way (which is good, because that's a difference of 16 million grass blades).