View space adjustments and basic clumping
Neither of these things are done particularly well. Clumping is very basic and fine for now. I don't believe my view space adjustment is working how I want it to and is just thickening everything regardless. I'm also pretty sure something may be off with normals or tangents on my grass blades because the specular highlights on the grass seems to be coming in a slightly different direction than the sun.